Add NpmBasedUrlUpdater
